pyproject.toml
[tool.poetry]
name = "python-web-crud-app-with-ddd"
version = "0.1.0"
description = ""
authors = ["heumsi <heumsi@naver.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.68.0"}
dependency-injector = "^4.35.2"
dataset = "^1.5.0"
python-ulid = "^1.0.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
requests = "^2.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
